Main API

DelaySSAToolkit.AbstractDSSAJumpAggregator

DelaySSAToolkit.DelayDirectJumpAggregation

DelaySSAToolkit.DelayJumpProblem

DelaySSAToolkit.DelayJumpProblem

DelaySSAToolkit.DelayJumpSet

DelaySSAToolkit.DelayMNRMJumpAggregation

DelaySSAToolkit.DelayRejectionJumpAggregation

DelaySSAToolkit.calculate_jump_rate

DelaySSAToolkit.compare_delay!

DelaySSAToolkit.create_Tstruct

DelaySSAToolkit.fill_cum_rates_and_sum!

DelaySSAToolkit.fill_rates_and_sum!

DelaySSAToolkit.find_next_delay_dt!

DelaySSAToolkit.find_next_delay_vec

DelaySSAToolkit.find_num_in_vec

DelaySSAToolkit.generate_delta!

DelaySSAToolkit.saveat_function!

DelaySSAToolkit.saveat_function_direct_method!

DelaySSAToolkit.update_delay_complete!

DelaySSAToolkit.update_delay_interrupt!

DelaySSAToolkit.update_delay_trigger!

SciMLBase.remake

DelaySSAToolkit.AbstractDSSAJumpAggregator — Type

An aggregator interface for Delay SSA-like algorithms.

Required Fields

next_jump # the next jump to execute

prev_jump # the previous jump that was executed

next_jump_time # the time of the next jump

end_time # the time to stop a simulation

cur_rates # vector of current propensity values

sum_rate # sum of current propensity values

ma_jumps # any MassActionJumps for the system (scalar form)

rates # vector of rate functions for ConstantRateJumps

affects! # vector of affect functions for ConstantRateJumps

save_positions # tuple for whether to save the jumps before and/or after event

rng # random number generator

next_delay # the index of the delay channel

num_next_delay # how many times needed for updating the states in the next_delay channel

time_to_next_jump # the time to the next jump (time gap)

dt_delay # the time to the next delay reaction

Optional fields:

dep_gr # dependency graph, dep_gr[i] = indices of reactions that should # be updated when rx i occurs.

DelaySSAToolkit.DelayDirectJumpAggregation — Type

#TODO

DelaySSAToolkit.DelayJumpProblem — Method

TODO

DelaySSAToolkit.DelayJumpProblem — Method

TODO

DelaySSAToolkit.DelayJumpSet — Type

-delay_trigger::Dict{Int,Any} # those reactions in the Markovian part that trigger the change of the state of the delay channels or/and the state of the reactants upon initiation. Keys: Indices of reactions defined in the Markovian part that can trigger the delay reactions; Values: Update functions(or Pair type) that decide how to update the delay channel or the state of the reactants.

-delay_interrupt::Dict{Int,Any} # those reactions in the Markovian part that change the state of the delay channels or/and the state of the reactants in the middle of on-going delay reactions. Keys: Indices of reactions defined in the Markovian part that can interrupt the delay reactions; Values: Update functions(or Pair type) that decide how to update the delay channel or the state of the reactants.

-delay_complete::Dict{Int,Any} # those reactions that are initiated by delay trigger reactions and change the state of the delay channels or/and the state of the reactants upon completion. Keys: Indices of the delay channel; Values: Update functions(or Pair type) that decide how to update the delay channel or the state of the reactants.

-delay_trigger_set::Vector{Int} # collect(keys(delay_trigger))

-delay_interrupt_set::Vector{Int} # collect(keys(delay_interrupt))

DelaySSAToolkit.DelayMNRMJumpAggregation — Type

DelaySSAToolkit.DelayRejectionJumpAggregation — Type

DelaySSAToolkit.calculate_jump_rate — Method

calculate_jump_rate(ma_jumps, rates, u, params, t, rx)

Recalculate the rate for the jump with index rx.

DelaySSAToolkit.compare_delay! — Method

function compare_delay!(p::AbstractDSSAJumpAggregator, de_chan, dt_delay, dt_reaction, t)

Compare dtdelay and dtreaction.

DelaySSAToolkit.create_Tstruct — Method

function create_Tstruct(de_chan::Vector{Vector{T}})

calculate Tstruct according to the de_chan.

DelaySSAToolkit.fill_cum_rates_and_sum! — Method

function fill_cum_rates_and_sum!(p::DelayDirectJumpAggregation, u, params, t)

This function modifies p.cur_rates and p.sum_rate

DelaySSAToolkit.fill_rates_and_sum! — Method

fill_rates_and_sum!(p::AbstractDSSAJumpAggregator, u, params, t)

Reevaluate all rates and their sum.

DelaySSAToolkit.find_next_delay_dt! — Method

function find_next_delay_dt!(p, integrator)

Find the minimal dt_delay in all delay channels.

DelaySSAToolkit.find_next_delay_vec — Method

find_next_delay_vec(A::Vector{Vector{T}}, x::T)

Returns two vectors. The first is the position index vector position_index of vector A, and the second is the vector num_in_vec composed of the number of values in the position index corresponding to position_index equal to x.

find the minimal dt_delay in various delay channel.

Examples

julia> A =  [[0.09,0.09,0.1],[0.3,0.09,0.1],[0.09]]
3-element Vector{Vector{Float64}}:
 [0.09, 0.09, 0.1]
 [0.3, 0.09, 0.1]
 [0.09]

 julia> x=0.09
 0.09

julia> find_next_delay_vec(A, x)
([1, 2, 3], [2, 1, 1])

DelaySSAToolkit.find_num_in_vec — Method

function find_num_in_vec(A::Vector{Vector{T}}, position_index::Vector{Int64}, x::T)

Find the number of values which in each vector elements equal to x according to the corresponding index position specified by the element in the position_index vector in the given vetcer A.

Examples

julia> A =  [[0.09,0.09,0.1],[0.3,0.09,0.1],[0.09]]
3-element Vector{Vector{Float64}}:
 [0.09, 0.09, 0.1]
 [0.3, 0.09, 0.1]
 [0.09]

julia> position_index =  [1,2,3]
3-element Vector{Int64}:
 1
 2
 3

julia> find_num_in_vec(A::Vector, position_index::Vector{Int64}, 0.09)
3-element Vector{Int64}:
2
1
1

DelaySSAToolkit.generate_delta! — Method

Create delta based on the shawdow variable u_shadow

DelaySSAToolkit.saveat_function! — Method

function saveat_function!(integrator, prev_t)

Note that this function does not change u and dechan, but only takes the snapshots of u and dechan accordingly

DelaySSAToolkit.saveat_function_direct_method! — Method

function saveat_function_direct_method!(integrator, prev_t)

Note that this function does not change u and dechan, but only takes the snapshots of u and dechan accordingly

DelaySSAToolkit.update_delay_complete! — Method

function update_delay_complete!(p, integrator)

This function modifies integrator.u and integrator.de_chan upon delay completion

DelaySSAToolkit.update_delay_interrupt! — Method

function update_delay_interrupt!(p, integrator)

DelaySSAToolkit.update_delay_trigger! — Method

function update_delay_trigger!(p, integrator)

SciMLBase.remake — Method

TODO for remaking