Main API

DelaySSAToolkit.DelayJumpSet — Type

mutable struct DelayJumpSet

A delay jump set...

Fields

delay_trigger

those reactions in the Markovian part that trigger the change of the state of the delay channels or/and the state of the reactants upon initiation.

delay_complete

those reactions in the Markovian part that change the state of the delay channels or/and the state of the reactants in the middle of on-going delay reactions.

delay_interrupt

those reactions that are initiated by delay trigger reactions and change the state of the delay channels or/and the state of the reactants upon completion.

delay_trigger_set

collect(keys(delay_trigger))

delay_interrupt_set

collect(keys(delay_interrupt))

Notes

delay_trigger::Dict{Int,Any}: those reactions in the Markovian part that trigger the change of the state of the delay channels or/and the state of the reactants upon initiation.

Keys: Indices of reactions defined in the Markovian part that can trigger the delay reactions;

Values: Update functions(or Pair type) that decide how to update the delay channel or the state of the reactants.

delay_interrupt::Dict{Int,Any}: those reactions in the Markovian part that change the state of the delay channels or/and the state of the reactants in the middle of on-going delay reactions.

Keys: Indices of reactions defined in the Markovian part that can interrupt the delay reactions;

Values: Update functions(or Pair type) that decide how to update the delay channel or the state of the reactants.

delay_complete::Dict{Int,Any}: those reactions that are initiated by delay trigger reactions and change the state of the delay channels or/and the state of the reactants upon completion.

Keys: Indices of the delay channel;

Values: Update functions(or Pair type) that decide how to update the delay channel or the state of the reactants.

delay_trigger_set::Vector{Int}: collect(keys(delay_trigger))

delay_interrupt_set::Vector{Int}: collect(keys(delay_interrupt))

DelaySSAToolkit.DelayJumpProblem — Type

function DelayJumpProblem(prob::DiscreteProblem, aggregator::AbstractDelayAggregatorAlgorithm, jumps::JumpSet, delayjumpsets::DelayJumpSet, de_chan0)

Fields

prob::DiscreteProblem

aggregator::AbstractDelayAggregatorAlgorithm

jumps::JumpSet

delayjumpsets::DelayJumpSet

de_chan0::Vector{Vector{T}}

TODO

function DelayJumpProblem(js::JumpSystem, prob, aggregator, delayjumpset, de_chan0; kwargs...)

Fields

js::JumpSystem: ...

prob::DiscreteProblem: ...

aggregator::AbstractDelayAggregatorAlgorithm

delayjumpsets::DelayJumpSet

de_chan0::Vector{Vector{T}}

TODO

More about defining a DelayJumpSet

Further notes on delay_trigger, delay_interrupt, delay_complete.

delay_trigger

delay_trigger defines a Dict type:

Keys: Indices of reactions defined in the Markovian part that can trigger the delay reaction;

Values: An update function or a Pair type that determines how to update the delay channel.

delay_interrupt

delay_complete

Types and Algorithms

DelaySSAToolkit.AbstractDelayAggregatorAlgorithm — Type

abstract type AbstractDelayAggregatorAlgorithm

An abstract type that contains delay stochastic simulation algorithms:

DelayDirect

DelayRejection

DelayMNRM

DelayDirectCR

DelaySSAToolkit.DelayDirect — Type

struct DelayDirect <: DelaySSAToolkit.AbstractDelayAggregatorAlgorithm

Delay Direct Method from Xiaodong Cai, "Exact stochastic simulation of coupled chemical reactions with delays", The Journal of Chemical Physics 126, 124108(2007).

DelaySSAToolkit.DelayRejection — Type

struct DelayRejection <: DelaySSAToolkit.AbstractDelayAggregatorAlgorithm

Delay Rejection Method from Barrio, Manuel, Kevin Burrage, André Leier, and Tianhai Tian. "Oscillatory regulation of Hes1: discrete stochastic delay modelling and simulation." PLoS computational biology 2, no. 9 (2006): e117..

DelaySSAToolkit.DelayMNRM — Type

struct DelayMNRM <: DelaySSAToolkit.AbstractDelayAggregatorAlgorithm

A modifed version of the Delay Next Reaction Method from David F. Anderson, "A modified Next Reaction Method for simulating chemical systems with time dependent propensities and delays", The Journal of Chemical Physics 128, 109903(2008).

DelaySSAToolkit.DelayDirectCR — Type

struct DelayDirectCR <: DelaySSAToolkit.AbstractDelayAggregatorAlgorithm

A modifed Composition-Rejection Direct Method with delays (DelayDirectCR), implementation combining features from the original article and from the code in DiffEqJump package: DirectCR : A constant-time kinetic Monte Carlo algorithm for simulation of large biochemical reaction networks, by A. Slepoy, A.P. Thompson and S.J. Plimpton, J. Chem. Phys, 128, 205101 (2008). and Efficient Formulations for Exact Stochastic Simulation of Chemical Systems, by S. Mauch and M. Stalzer, ACM Trans. Comp. Biol. and Bioinf., 8, No. 1, 27-35 (2010).