Main API

DelaySSAToolkit.AbstractDSSAJumpAggregator

DelaySSAToolkit.ChannelSolution

DelaySSAToolkit.DSSAIntegrator

DelaySSAToolkit.DelayDirectJumpAggregation

DelaySSAToolkit.DelayJumpSet

DelaySSAToolkit.DelayMNRMJumpAggregation

DelaySSAToolkit.DelayRejection

DelaySSAToolkit.DelayRejectionJumpAggregation

DelaySSAToolkit.calculate_jump_rate

DelaySSAToolkit.compare_delay!

DelaySSAToolkit.create_Tstruct

DelaySSAToolkit.fill_rates_and_sum!

DelaySSAToolkit.find_next_delay_vec

DelaySSAToolkit.find_num_in_vec

DelaySSAToolkit.generate_delta!

DelaySSAToolkit.update_delay_channel!

DelaySSAToolkit.update_delay_complete!

DelaySSAToolkit.update_state_delay_Direct!

DelaySSAToolkit.AbstractDSSAJumpAggregator — Type

An aggregator interface for SSA-like algorithms.

Required Fields

next_jump # the next jump to execute

prev_jump # the previous jump that was executed

next_jump_time # the time of the next jump

end_time # the time to stop a simulation

cur_rates # vector of current propensity values

sum_rate # sum of current propensity values

ma_jumps # any MassActionJumps for the system (scalar form)

rates # vector of rate functions for ConstantRateJumps

affects! # vector of affect functions for ConstantRateJumps

save_positions # tuple for whether to save the jumps before and/or after event

rng # random number generator

Optional fields:

dep_gr # dependency graph, dep_gr[i] = indices of reactions that should # be updated when rx i occurs.

DelaySSAToolkit.ChannelSolution — Type

-delaytrigger: Dict 记录哪些 reactions会引发一个delay反应：keys：reaction的idx，values Dict 记录在第i个delaychannel 增加一个列向量 为delay的时间. 这里包含了两种情况，1.在delay前对状态改变(pre-delay)，这个是由reaction来完成的；2.在delay后对状体改变(post-delay)，这个是由delayvec中完成的 -delayinterrupt: Dict reactions 能够对 delay channel中造成影响的 keys : reaction idx -> returns a Function : how the molecules in a channel or multiple channels will be consumed -delay_complete: Dict keys:第 i 个delay channel values 完成后会引起一个 (post-delay) state-update, values: stoichiometric vector 长度是反应物长度

DelaySSAToolkit.DSSAIntegrator — Type

dechan: 记录delaychannel, Dict(), keys 记录哪一个delaychannel, values 记录在这个delaychannel中所等待的时间列表

prevdechan: For saveat update

DelaySSAToolkit.DelayDirectJumpAggregation — Type

#TODO

DelaySSAToolkit.DelayJumpSet — Type

-delaytrigger: Dict 记录哪些 reactions会引发一个delay反应：keys：reaction的idx, values: delaytrigger_affect!

-delay_complete: Dict keys:第 i 个delay channel values 完成后会引起一个 (post-delay) state-update, values: stoichiometric vector 长度是反应物长度

-delay_interrupt: Dict reactions 能够对 delay channel中造成影响的 keys : reaction idx -> returns a Function : how the molecules in a channel or multiple channels will be consumed

delaytriggerset: 把所有会引发delay 的 reactions 的 idx 收集起来

delayinterruptset: 把所有会引发delay channel 改变的 reactions 的 idx 收集起来

DelaySSAToolkit.DelayMNRMJumpAggregation — Type

next_delay: the idx of delay channel

numnextdelay: how many times needed for updating the states in the next_delay channel

DelaySSAToolkit.DelayRejection — Type

Gillespie, Daniel T. (1976). A General Method for Numerically Simulating the Stochastic Time Evolution of Coupled Chemical Reactions. Journal of Computational Physics. 22 (4): 403–434. doi:10.1016/0021-9991(76)90041-3.

DelaySSAToolkit.DelayRejectionJumpAggregation — Type

next_delay: the idx of delay channel

numnextdelay: how many times needed for updating the states in the next_delay channel

DelaySSAToolkit.calculate_jump_rate — Method

calculate_jump_rate(ma_jumps, rates, u, params, t, rx)

Recalculate the rate for the jump with index rx.

DelaySSAToolkit.compare_delay! — Method

Compare delay dt with reaction dt

DelaySSAToolkit.create_Tstruct — Method

function create_Tstruct(de_chan::Vector{Vector{T}})

calculate Tstruct according to the de_chan

Arguments

de_chan::Vector{Vector{T}}: where T is the tpye of t

DelaySSAToolkit.fill_rates_and_sum! — Method

fill_rates_and_sum!(p::AbstractDSSAJumpAggregator, u, params, t)

Reevaluate all rates and their sum.

DelaySSAToolkit.find_next_delay_vec — Method

find_next_delay_vec(A::Vector{Vector{T}}, x::T)

Returns two vectors. The first is the position index vector position_index of vector A, and the second is the vector num_in_vec composed of the number of values in the position index corresponding to position_index equal to x.

find the minimal dt_delay in various delay channel.

Examples

julia> A =  [[0.09,0.09,0.1],[0.3,0.09,0.1],[0.09]]
3-element Vector{Vector{Float64}}:
 [0.09, 0.09, 0.1]
 [0.3, 0.09, 0.1]
 [0.09]

 julia> x=0.09
 0.09

julia> find_next_delay_vec(A, x)
([1, 2, 3], [2, 1, 1])

DelaySSAToolkit.find_num_in_vec — Method

function find_num_in_vec(A::Vector{Vector{T}}, position_index::Vector{Int64}, x::T)

Find the number of values which in each vector elements equal to x according to the corresponding index position specified by the element in the position_index vector in the given vetcer A.

Examples

julia> A =  [[0.09,0.09,0.1],[0.3,0.09,0.1],[0.09]]
3-element Vector{Vector{Float64}}:
 [0.09, 0.09, 0.1]
 [0.3, 0.09, 0.1]
 [0.09]

julia> position_index =  [1,2,3]
3-element Vector{Int64}:
 1
 2
 3

julia> find_num_in_vec(A::Vector, position_index::Vector{Int64}, 0.09)
3-element Vector{Int64}:
2
1
1

DelaySSAToolkit.generate_delta! — Method

Create delta based on the shawdow variable u_shadow

DelaySSAToolkit.update_delay_channel! — Method

Update the delay channel

DelaySSAToolkit.update_delay_complete! — Method

Update the state upon delay completion

DelaySSAToolkit.update_state_delay_Direct! — Method

Update state according up the next_jump;